CA-2200 - Multi route support

+++ samples/src/gpsTracking/route.ts
@@ -1,10 +1,12 @@
 import { SDK } from '@crewdle/web-sdk';
 import { vendorId, accessToken } from '../credentials';
-import { route1, route2, route3 } from './routes';
+import { route1, route2, route3, route4 } from './routes';
 import { calculateLength, IRouteDetails } from './helpers';
+import { map } from 'leaflet';
 
 interface ITrackingMessage {
   coordinate: number[];
+  oldPosition: number[];
   routeDetails: IRouteDetails;
 }
 
@@ -18,8 +20,10 @@ let usedRoute: {
 };
 let lastPosition: number[];
 let currentLength = 0;
+let mapDetails: IRouteDetails;
+let fullLength = 0;
 
-export async function start(clusterId: string, userId: string) {
+export async function start(clusterId: string, userId: string, mapType: string) {
   userIdFrame = userId;
 
   if (userIdFrame === 'route1') {
@@ -28,6 +32,15 @@ export async function start(clusterId: string, userId: string) {
     usedRoute = route2;
   } else if (userIdFrame === 'route3') {
     usedRoute = route3;
+  }
+
+  if (mapType === 'user-map') {
+    mapDetails = {
+      length: usedRoute.length + route4.length,
+      start: usedRoute.start,
+      end: route4.end,
+      percentage: 0,
+    };
   }
 
   const sdk = await SDK.getInstance(vendorId, accessToken, {
@@ -43,9 +56,13 @@ export async function start(clusterId: string, userId: string) {
   const cluster = await sdk.joinCluster(clusterId)
   const pubsub = cluster.openPubSubTopic<ITrackingMessage>('tracking');
 
-  const usedCoordinates = usedRoute.coordinates;
+  let usedCoordinates = usedRoute.coordinates;
 
-  const fullLength = usedRoute.length;
+  if (mapType === 'user-map') {
+    fullLength = usedRoute.length + route4.length;
+  } else {
+    fullLength = usedRoute.length;
+  }
   lastPosition = usedCoordinates[0];
 
   const workerScript = `
@@ -61,23 +78,36 @@ export async function start(clusterId: string, userId: string) {
   const worker = new Worker(URL.createObjectURL(blob));
 
   worker.onmessage = function(e) {
+    if (mapType === 'user-map') {
+      usedRoute.length = mapDetails.length;
+      usedRoute.start = mapDetails.start;
+      usedRoute.end = mapDetails.end;
+    }
     let percentage = 0;
     if (currentIndex >= usedCoordinates.length) {
-      worker.terminate();
-      pubsub.publish({ coordinate: usedCoordinates[usedCoordinates.length - 1], routeDetails: {
-        length: usedRoute.length,
-        start: usedRoute.start,
-        end: usedRoute.end,
-        percentage: 100,
-      }});
-      return;
+      if (mapType === "user-map" && usedRoute !== route4) {
+        currentIndex = 0;
+        usedRoute = route4;
+        usedCoordinates = usedRoute.coordinates;
+        lastPosition = usedRoute.coordinates[0];
+        worker.postMessage("start");
+      } else {
+        worker.terminate();
+        pubsub.publish({ coordinate: usedCoordinates[usedCoordinates.length - 1], oldPosition: lastPosition, routeDetails: {
+            length: usedRoute.length,
+            start: usedRoute.start,
+            end: usedRoute.end,
+            percentage: 100,
+          }});
+        return;
+      }
     }
 
     const data = usedCoordinates[currentIndex];
     currentLength += calculateLength([lastPosition, data]);
     percentage = Math.abs(currentLength / fullLength) * 100;
 
-    pubsub.publish({ coordinate: data, routeDetails: {
+    pubsub.publish({ coordinate: data, oldPosition: lastPosition, routeDetails: {
       length: usedRoute.length,
       start: usedRoute.start,
       end: usedRoute.end,
